requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.2.0
umap-learn>=0.5.0
hdbscan>=0.8.0
faiss-cpu>=1.7.0
httpx>=0.24.0
python-multipart>=0.0.6
pyarrow>=10.0.0
sqlalchemy>=2.0.0